refactor(api): isolate bot availability check
